+++ hist2query/utils.py
@@ -5,6 +5,14 @@ def numerical_or_none(arg: Union[str, int, None],
     """
     Set an argparse arg as either None or a numerical type
     """
-    if str(arg).lower() in ("none", "null"):
+    if str(arg).lower() in ("none", "null") or arg is None:
         return None
     return numerical_type(arg)
+
+def str_or_none(arg: Union[str, int, None]=None):
+    """
+    Set an argparse arg as either None or a numerical type
+    """
+    if str(arg).lower() in ("none", "null") or arg is None:
+        return None
+    return str(arg)